// src/app/Models/Drug.ts
export class Drug {
    // ID:number;
    code: string;
    tradeName: string;
    genericName: string;
    strength:number;
    pack:number;
    license: string;
    img: string;
    reOrderLevel: string;
    barCode: string;
    TheraSubGroupID:number;
    FormID:number;
    FirmID:number;
    UnitID:number;
    ROADID:number;
    CountryID:number;
  }